backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
openai==1.3.7
stripe==7.8.0
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.25.2
reportlab==4.0.7
jinja2==3.1.2
aiofiles==23.2.1